pyproject.toml
[tool.poetry]
name = "nobu-cli"
version = "1.0.0"
description = ""
authors = ["Guilherme Marim <3b853b@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.6.1"
requests = "^2.32.3"
rich = "^13.9.4"
requests-cache = "^1.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
taskipy = "^1.14.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
bandit = "^1.7.10"

[tool.poetry.scripts]
nobu = "nobu.app:run"

[tool.pytest.ini_options]
pythonpath = '.'

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLC0415', 'PLR6301']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
format = 'ruff check . --fix && ruff format .'
pre_test = 'task format'
test = 'pytest -s -x --cov=nobu -vv'
post_test = 'coverage html'
pre_docker_run = 'docker build -t nobu:latest .'
docker_run = 'docker run -it --rm nobu:latest'
post_docker_run = 'docker rmi nobu:latest'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
